Commit 1
I added on line 20 card1 to group2Members with the identifier 'Card1'.
I added on line 21 card2 to group2Members with the identifier 'Card2'.
I added on line 22 card3 to group2Members with the identifier 'Card3'.
I added on line 23 group2Members (the card container) to the main JFrame group2.
I set on line 24 group2 visible so the window and cards appear on the screen.